trying to fix tests and github actions also decided that it should be one deployment per repo and will refactor after tests work
