+++ fastapi-crud/entrypoint.sh
@@ -1,0 +1,1 @@
+python -m src
