+++ fastapi-crud/requirements.txt
@@ -1,16 +1,26 @@
+aiocache==0.11.1
+aiohttp==3.8.3
+aiosignal==1.3.1
 anyio==3.7.0
+async-timeout==4.0.2
 asyncpg==0.27.0
 attrs==22.1.0
+bcrypt==4.0.1
 blinker==1.6.2
 certifi==2023.5.7
+cffi==1.15.1
 cfgv==3.3.1
+charset-normalizer==2.1.1
 click==8.1.3
+cryptography==41.0.1
 distlib==0.3.5
+ecdsa==0.18.0
 exceptiongroup==1.1.1
 fastapi==0.98.0
 filelock==3.8.0
 flake8==5.0.4
 Flask==2.3.2
+frozenlist==1.3.3
 greenlet==2.0.2
 h11==0.14.0
 httpcore==0.17.2
@@ -26,17 +36,25 @@ mccabe==0.7.0
 multidict==6.0.4
 nodeenv==1.7.0
 packaging==21.3
+passlib==1.7.4
 platformdirs==2.5.2
 pluggy==1.0.0
 pre-commit==2.20.0
 py==1.11.0
+pyasn1==0.5.0
 pycodestyle==2.9.1
+pycparser==2.21
 pydantic==1.10.9
 pyflakes==2.5.0
 pyparsing==3.0.9
-pytest==7.1.2
+pytest==7.4.0
+pytest-env==0.8.2
 python-dotenv==1.0.0
+python-jose==3.3.0
+python-multipart==0.0.6
 PyYAML==6.0
+rsa==4.9
+six==1.16.0
 sniffio==1.3.0
 SQLAlchemy==2.0.17
 starlette==0.27.0
