+++ fastapi-crud/src/tests/test_template.py
